fix(memory,agent): fix postgres compile failure and skill provider startup init
- #4078: gate duplicate checkpoint_wal and community_ids_sql definitions in
  graph/store/mod.rs with #[cfg(all(feature = "sqlite", not(feature = "postgres")))]
  to avoid E0428/E0592 when both features are active; add postgres-gated
  implementations for load_graph and list_graphs using sqlx::query::<Postgres>()
  turbofish; add postgres equivalents for five SQLite-dialect methods in
  store/skills.rs (datetime/INSTR/max → NOW()-INTERVAL/strpos/GREATEST); fix
  u32→i64 bind params across corrections.rs, experiments.rs, messages/mod.rs,
  preferences.rs, response_cache.rs, plan_cache.rs

- #4083: add with_skill_provider_names() builder method in agent/builder.rs and
  wire it in runner.rs and daemon.rs so generation_provider_name and
  disambiguate_provider_name are populated from config at startup, not only
  on hot-reload

Closes #4078
Closes #4083

+++ crates/zeph-memory/src/response_cache.rs
@@ -111,7 +111,7 @@ impl ResponseCache {
         ))
         .bind(embedding_model)
         .bind(now)
-        .bind(max_candidates)
+        .bind(i64::from(max_candidates))
         .fetch_all(&self.pool)
         .await?;
